requirements.txt
python-dotenv
requests
compress_json